Add files via upload
DEA RUSSELL 2000:
This idea was taken from web which gives free trading strategies ideas, specifically the "quantitative" ones, the link which I get the idea is the following one: https://www.quantifiedstrategies.com/quantitative-trading-strategies/
The strategy says the following:
Buy on the close of the first trading day after the 23rd of June. Sell on the close on the first trading day of July.